Account Manager is now working, the user needs to add a account before use the application, after the first account been add, the user no longer needs to login.
There is a splash for showing the logo for about 3s every time the application is started.

Fixed some issues about screen rotation, still need more tests.

The name of the place is now passed by bundle to the PlaceActivity

Still need to figure out how to fix the NPE in LocationManager when there is no lastKnowPlace.